requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
discord.py==2.6.2
mcp==1.13.1
notion_client==2.5.0
playwright==1.55.0
playwright_stealth==2.0.0
pydantic==2.11.7
python-dotenv==1.1.1
redis==6.4.0
Requests==2.32.5
celery==5.4.0
celery-redbeat==2.2.0